refactor(platform-wallet): move mnemonic recover-flow logic out of the FFI crate
Per review feedback (FFI crates should hold no logic, only C-type
marshalling): move the recover-flow algorithm into a new non-FFI module
platform_wallet::mnemonic -- CJK ideographic auto-split + punctuation cleanup
(incl. the NFC fix), the 7-language DashSync bundled-language union, and the
BCP-47 language mapping. The four extern "C" entry points now just marshal C
strings and delegate; normalize stays a direct key_wallet::Mnemonic call (a
raw BIP-39 primitive, no platform policy). Adds an input length cap to
cleanup_phrase so a huge no-space CJK paste can't drive a superlinear scan.

Tests: the recover-flow-logic tests move to platform_wallet::mnemonic; the
ones that merely re-tested key_wallet behaviour through wrappers are deleted
(key_wallet already covers normalize/validate/membership/Trezor-vectors); the
FFI crate keeps only ffi_roundtrip (the extern "C"/CString/null-pointer
boundary). Per-language validate coverage now also lives natively in
key-wallet (dashpay/rust-dashcore#807).

The extern "C" signatures are unchanged, so the cbindgen header is identical
and the Swift SDK needs no change. unicode-normalization moves from the FFI
crate to platform-wallet (now the only direct user).
