+++ packages/rs-platform-wallet/Cargo.toml
@@ -48,6 +48,11 @@ image = { version = "0.25", default-features = false, features = ["png", "jpeg",
 # Security
 zeroize = "1"
 
+# Unicode normalization for the recover-flow mnemonic cleanup (NFKD +
+# combining-mark filtering in `mnemonic::cleanup_phrase`). Previously a direct
+# dep of rs-platform-wallet-ffi, before that logic moved into this crate.
+unicode-normalization = "0.1"
+
 # Shielded pool (optional, behind `shielded` feature)
 grovedb-commitment-tree = { git = "https://github.com/dashpay/grovedb", rev = "fc814983d4d36c6ea049642556b9a31ab8d4dfaa", optional = true }
 # Direct `rusqlite` access so `FileBackedShieldedStore::open_path` can set
