+++ packages/rs-platform-wallet/src/lib.rs
@@ -18,6 +18,7 @@ pub mod changeset;
 pub mod error;
 pub mod events;
 pub mod manager;
+pub mod mnemonic;
 pub mod spv;
 pub mod wallet;
